requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-pinecone
langchain-groq
pinecone
sentence-transformers
pypdf